Refactor: 보고서 상세 경로 공통화 및 컴포넌트 폴더 구조 정리
